Day 13: Simplify map reading
We can split rows into chars, and then just directly transpose that.
This is especially useful in part 2, as pre-splitting into chars means
we don't have to do that on each loop in `get_reflection_point`. It also
makes it easier to have `maps` just be the rows, rather than an object
containing the rows _and_ columns representation.